+++ src/main/java/com/github/PatrykKukula/Photovoltaic/materials/calculator/Dto/Installation/InstallationUpdateDto.java
@@ -28,7 +28,7 @@ public class InstallationUpdateDto implements InstallationInterface {
     @NotNull(message = "Pick phase number")
     private PhaseNumber phaseNumber;
     @NotNull(message = "Specify number of DC strings")
-    private Long strings;
+    private Integer strings;
     private List<RowDto> rows;
     @NotNull(message = "Choose whether object has lighting protection installation")
     private boolean lightingProtection;
